requirements.txt
robotframework
playwright
robotframework-browser
allure-python-commons
